+++ spec/whois/record/parser/responses/whois.nic.tr/status_available_spec.rb
@@ -21,6 +21,21 @@ describe Whois::Record::Parser::WhoisNicTr, "status_available.expected" do
     described_class.new(part)
   end
 
+  describe "#disclaimer" do
+    it do
+      lambda { subject.disclaimer }.should raise_error(Whois::AttributeNotSupported)
+    end
+  end
+  describe "#domain" do
+    it do
+      lambda { subject.domain }.should raise_error(Whois::AttributeNotSupported)
+    end
+  end
+  describe "#domain_id" do
+    it do
+      lambda { subject.domain_id }.should raise_error(Whois::AttributeNotSupported)
+    end
+  end
   describe "#status" do
     it do
       subject.status.should == :available
@@ -49,6 +64,29 @@ describe Whois::Record::Parser::WhoisNicTr, "status_available.expected" do
   describe "#expires_on" do
     it do
       subject.expires_on.should == nil
+    end
+  end
+  describe "#registrar" do
+    it do
+      lambda { subject.registrar }.should raise_error(Whois::AttributeNotSupported)
+    end
+  end
+  describe "#registrant_contacts" do
+    it do
+      subject.registrant_contacts.should be_a(Array)
+      subject.registrant_contacts.should == []
+    end
+  end
+  describe "#admin_contacts" do
+    it do
+      subject.admin_contacts.should be_a(Array)
+      subject.admin_contacts.should == []
+    end
+  end
+  describe "#technical_contacts" do
+    it do
+      subject.technical_contacts.should be_a(Array)
+      subject.technical_contacts.should == []
     end
   end
   describe "#nameservers" do