+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 # Master
 
+- NEW: Added full whois.nic.tr parser.
+
 - CHANGED: Updated whois.nic.hn parsers to the new response format.
 
 - CHANGED: Updated whois.nic.as parsers to the new response format.